firmware_type: Refactored the FirmwareType enum and logic into its own file
